add: image slide component and some other adustments
